Move signals out of models per latest Django best practice
