+++ symposion/proposals/apps.py
@@ -1,0 +1,9 @@
+from django.apps import AppConfig
+
+
+class SymposionProposalsConfig(AppConfig):
+    name = 'symposion.proposals'
+
+    def ready(self):
+        # Hook up signals now that models are done loading
+        import symposion.proposals.signals
